Agents/D-Agents/Agent-2/requirements.txt
# Core FastAPI dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# Google Cloud and AI dependencies
google-generativeai>=0.4.0
google-cloud-firestore>=2.14.0
google-cloud-aiplatform>=1.43.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Core dependencies
requests>=2.31.0
python-multipart>=0.0.7
PyJWT>=2.8.0
cryptography>=42.0.0

# Image processing (compatible with Python 3.12)
Pillow>=10.2.0

# Numpy - use a version compatible with Python 3.12
numpy>=1.26.0

# Additional utilities
python-dotenv>=1.0.0

# MongoDB dependencies
pymongo>=4.6.0

# WebSocket dependencies
websockets>=12.0
